Example/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
deepface==0.0.79
google-generativeai==0.3.2
opencv-python-headless==4.8.1.78
numpy==1.24.3
gunicorn==21.2.0
tensorflow==2.13.0
requests==2.31.0
